requirements.txt
# Production
Flask
python-dotenv
Flask-SocketIO
google-generativeai
pypandoc
PyMuPDF

# Testing
pytest
python-socketio
